requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
chromadb
cachetools
tenacity
slowapi
python-json-logger
prometheus-fastapi-instrumentator